refactor(chrome): one icon button with a hover label for bar and navigator
The bar's icon button moves to ChromeIconButton so the navigator's zoom,
fit, lock and minimap controls get the same hover label and states. Native
title tooltips come off the buttons that carry the custom label, so they
no longer show twice.
